requirement.txt
Flask==2.0.1
Flask-Cors==3.0.10
fitz==0.0.1.dev2
opencv-python==4.5.3.56
langchain
langchain-extensions
PILLOW==8.2.0
pdfminer.six==20201018
easyocr==1.3.2
transformers==4.11.3
ultralytics==8.0.20
google-generativeai
langchain-google-genai
python-dotenv
